Rearranged instances of class Imovel
